Add XElementMediaTypeHandlerBase
