+++ src/Simple.Web.Xml/ExplicitXmlMediaTypeHandler.cs
@@ -1,32 +1,16 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.IO;
 using System.Threading.Tasks;
 using System.Xml.Linq;
-using Inflector;
 using Simple.Web.DependencyInjection;
 using Simple.Web.Helpers;
-using Simple.Web.Links;
 using Simple.Web.MediaTypeHandling;
 
 namespace Simple.Web.Xml
 {
     [MediaTypes(MediaType.Xml, "application/*+xml")]
-    public class ExplicitXmlMediaTypeHandler : MediaTypeHandlerBase<XElement>
+    public class ExplicitXmlMediaTypeHandler : XElementMediaTypeHandlerBase
     {
         private object _outputConverter;
-
-        protected override void AddWireFormattedLinks(XElement wireFormattedItem, IEnumerable<Link> itemLinks)
-        {
-            if (itemLinks == null)
-            {
-                return;
-            }
-            foreach (Link link in itemLinks)
-            {
-                wireFormattedItem.Add(link.ToXml());
-            }
-        }
 
         protected override Task<T> FromWireFormat<T>(XElement wireFormat)
         {
@@ -36,11 +20,6 @@ namespace Simple.Web.Xml
                 converter = container.Get<XmlConverter<T>>();
             }
             return TaskHelper.Completed(converter.FromXml(wireFormat));
-        }
-
-        protected override Task<XElement> ReadInput(Stream inputStream)
-        {
-            return TaskHelper.Completed(XElement.Load(inputStream));
         }
 
         protected override XElement ToWireFormat(object item)
@@ -76,24 +55,6 @@ namespace Simple.Web.Xml
                 converter = (XmlConverter<T>) _outputConverter;
             }
             return converter.ToXml(item);
-        }
-
-        protected override XElement WrapCollection(IList<XElement> collection, IEnumerable<Link> collectionLinks)
-        {
-            //todo should we wrap a single item?
-            var xml = new XElement(collection[0].Name.LocalName.Pluralize());
-            foreach (XElement element in collection)
-            {
-                xml.Add(element);
-            }
-            //todo add collection links?
-            return xml;
-        }
-
-        protected override Task WriteOutput(XElement output, Stream outputStream)
-        {
-            output.Save(outputStream);
-            return TaskHelper.Completed();
         }
     }
 }
